refactor: cookie get, set 함수 utils 하위로 이동
- pr 피드백 반영
