+++ src/common/utils/cookie/type.ts
